+++ udata_front/theme/gouvfr/assets/js/composables/useCollapse.ts
@@ -1,4 +1,4 @@
-import { MaybeRefOrGetter, Ref, ref, toValue } from 'vue'
+import { MaybeRefOrGetter, ref, toValue } from 'vue'
 import useKeyCodes from "./useKeyCodes";
 import { getRandomId } from '@datagouv/components/ts';