+++ udata_front/theme/gouvfr/datagouv-components/src/api/index.ts
@@ -1,5 +1,6 @@
 export * from "./api";
 export * from "./frequency";
 export * from "./granularity";
+export * from "./reuses";
 export * from "./schemas";
 export * from "./zones";