+++ udata_front/theme/gouvfr/datagouv-components/src/index.ts
@@ -6,8 +6,8 @@ import type { Granularity, Granularities } from "./types/granularity";
 import type { Harvest } from "./types/harvest";
 import type { License } from "./types/licenses";
 import type { Member, MemberRole, NewOrganization, Organization } from "./types/organizations";
-import type { Owned } from "./types/owned";
-import type { Reuse } from "./types/reuses";
+import type { Owned, OwnedWithId } from "./types/owned";
+import type { NewReuse, Reuse, ReuseType } from "./types/reuses";
 import type { CommunityResource, FileResourceFileType, RemoteResourceFileType, ResourceFileType, ResourceType, Resource } from "./types/resources";
 import type { Weight, WellType } from "./types/ui";
 import type { User } from "./types/users";
@@ -32,8 +32,10 @@ export type {
   MemberRole,
   NewDataset,
   NewOrganization,
+  NewReuse,
   Organization,
   Owned,
+  OwnedWithId,
   Quality,
   Rel,
   RemoteResourceFileType,
@@ -41,6 +43,7 @@ export type {
   ResourceFileType,
   ResourceType,
   Reuse,
+  ReuseType,
   User,
   Weight,
   WellType,